+++ mwparserfromhell/utils.py
@@ -45,14 +45,7 @@ def parse_anything(value, context=0, skip_style_tags=False):
     :py:class:`~.Template`, such as :py:meth:`wikicode.insert()
     <.Wikicode.insert>` or setting :py:meth:`template.name <.Template.name>`.
 
-    If given, *context* will be passed as a starting context to the parser.
-    This is helpful when this function is used inside node attribute setters.
-    For example, :py:class:`~.ExternalLink`\ 's :py:attr:`~.ExternalLink.url`
-    setter sets *context* to :py:mod:`contexts.EXT_LINK_URI <.contexts>` to
-    prevent the URL itself from becoming an :py:class:`~.ExternalLink`.
-
-    If *skip_style_tags* is ``True``, then ``''`` and ``'''`` will not be
-    parsed, but instead will be treated as plain text.
+    Additional arguments are passed directly to :py:meth:`.Parser.parse`.
     """
     from .parser import Parser
     from .wikicode import Wikicode